# examples/axum-example/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/complex-struct-body": {
      "post": {
        "operationId": "mod_file_with_complex_struct_body",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplexStructBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/complex-struct-body-with-rename": {
      "post": {
        "operationId": "mod_file_with_complex_struct_body_with_rename",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplexStructBodyWithRename"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/map-query": {
      "get": {
        "operationId": "mod_file_with_map_query",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "age",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "uint32"
            }
          },
          {
            "name": "optional_age",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "uint32",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/path/multi-path/{arg}/{var1}/{var2}": {
      "get": {
        "operationId": "mod_file_with_multi_path",
        "parameters": [
          {
            "name": "arg",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "var1",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "var2",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/path/multi-path/{var1}": {
      "get": {
        "operationId": "mod_file_with_multi_path_single",
        "parameters": [
          {
            "name": "var1",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/path/multi-path2/{arg}/{var1}/{var2}": {
      "get": {
        "operationId": "mod_file_with_multi_path_2",
        "parameters": [
          {
            "name": "arg",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "var1",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "var2",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/struct-body": {
      "post": {
        "operationId": "mod_file_with_struct_body",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StructBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/struct-body-with-optional": {
      "post": {
        "operationId": "mod_file_with_struct_body_with_optional",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StructBodyWithOptional"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/struct-query": {
      "get": {
        "operationId": "mod_file_with_struct_query",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "age",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "uint32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/terms": {
      "get": {
        "operationId": "list_terms",
        "tags": [
          "terms"
        ],
        "parameters": [
          {
            "name": "termsType",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TermsType"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TermsQuery"
                }
              }
            }
          }
        }
      }
    },
    "/test-struct": {
      "get": {
        "operationId": "mod_file_with_test_struct",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "age",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "uint32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestStruct"
                }
              }
            }
          }
        }
      }
    },
...
    "/typed-header": {
      "get": {
        "operationId": "typed_header_jwt",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "typed_header",
        "parameters": [
          {
            "name": "user-agent",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "content-type",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComplexStructBody": {
        "type": "object",
        "properties": {
          "age": {
            "type": "integer",
            "format": "uint32"
          },
          "array": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "map": {
            "type": "object",
            "properties": {},
            "required": [],
            "additionalProperties": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "nested_array": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StructBodyWithOptional"
            }
          },
          "nested_map": {
            "type": "object",
            "properties": {},
            "required": [],
            "additionalProperties": {
              "$ref": "#/components/schemas/StructBodyWithOptional"
            }
          },
          "nested_struct": {
            "$ref": "#/components/schemas/StructBodyWithOptional"
          },
          "nested_struct_array": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StructBodyWithOptional"
            }
          },
          "nested_struct_array_map": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {},
              "required": [],
              "additionalProperties": {
                "$ref": "#/components/schemas/StructBodyWithOptional"
              }
            }
          },
          "nested_struct_map": {
            "type": "object",
            "properties": {},
            "required": [],
            "additionalProperties": {
              "$ref": "#/components/schemas/StructBodyWithOptional"
            }
          },
          "nested_struct_map_array": {
            "type": "object",
            "properties": {},
            "required": [],
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/StructBodyWithOptional"
              },
              "type": "array"
            }
          }
        },
        "required": [
          "name",
          "age",
          "nested_struct",
          "array",
          "map",
          "nested_array",
          "nested_map",
          "nested_struct_array",
          "nested_struct_map",
          "nested_struct_array_map",
          "nested_struct_map_array"
        ]
      },
      "ComplexStructBodyWithRename": {
        "type": "object",
        "properties": {
          "age": {
            "type": "integer",
            "format": "uint32"
          },
          "array": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "map": {
            "type": "object",
            "properties": {},
            "required": [],
            "additionalProperties": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "nestedArray": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StructBodyWithOptional"
            }
          },
          "nestedMap": {
            "type": "object",
            "properties": {},
            "required": [],
            "additionalProperties": {
              "$ref": "#/components/schemas/StructBodyWithOptional"
            }
          },
          "nestedStruct": {
            "$ref": "#/components/schemas/StructBodyWithOptional"
          },
          "nestedStructArray": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StructBodyWithOptional"
            }
          },
          "nestedStructArrayMap": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {},
              "required": [],
              "additionalProperties": {
                "$ref": "#/components/schemas/StructBodyWithOptional"
              }
            }
          },
          "nestedStructMap": {
            "type": "object",
            "properties": {},
            "required": [],
            "additionalProperties": {
              "$ref": "#/components/schemas/StructBodyWithOptional"
            }
          },
          "nestedStructMapArray": {
            "type": "object",
            "properties": {},
            "required": [],
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/StructBodyWithOptional"
              },
              "type": "array"
            }
          }
        },
        "required": [
          "name",
          "age",
          "nestedStruct",
          "array",
          "map",
          "nestedArray",
          "nestedMap",
          "nestedStructArray",
          "nestedStructMap",
          "nestedStructArrayMap",
          "nestedStructMapArray"
        ]
      },
...
      "StructBody": {
        "type": "object",
        "properties": {
          "age": {
            "type": "integer",
            "format": "uint32"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "age"
        ]
      },
      "StructBodyWithOptional": {
        "type": "object",
        "properties": {
          "age": {
            "type": "integer",
            "format": "uint32",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "TermsQuery": {
        "type": "object",
        "properties": {
          "termsType": {
            "$ref": "#/components/schemas/TermsType"
          }
        },
        "required": [
          "termsType"
        ]
      },
      "TermsType": {
        "type": "string",
        "enum": [
          "terms",
          "privacy"
        ]
      },
      "TestStruct": {
        "type": "object",
        "properties": {
          "age": {
            "type": "integer",
            "format": "uint32"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "age"
        ]
      },
...
    }
  },
...
}